requirements.txt
django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
drf-yasg